refactor(files): restructure folders and files
