+++ src/components/DatePicker/index.ts
@@ -1,0 +1,3 @@
+import DatePicker from './DatePickerView';
+
+export default DatePicker;
